requirements.txt
# ============================================================
# 两阶段集成框架依赖 — Python 3.10+
# ============================================================

# 数值计算
numpy>=1.24,<2.0
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
statsmodels>=0.14
pmdarima>=2.0.4

# 深度学习
torch>=2.0
torchsde>=0.2.5

# 混沌 / 动力学特征
nolds>=0.5.2
antropy>=0.1.6
pyinform>=0.2.0

# 可解释性
shap>=0.44

# 绘图
matplotlib>=3.7

# Excel
openpyxl>=3.1

# 工具
tqdm>=4.66
python-dotenv>=1.0
